requirements.txt
arango==0.2.1
click==8.1.3
numpy
pandas
hydra-core
hydra
pyArango==2.0.1
pytest==7.2.0
python-dotenv==0.21.0
python_arango==7.5.5
PyYAML==6.0
requests==2.28.1
sentence_transformers==2.2.2
setuptools==65.7.0
torch_geometric==2.2.0
torch-tb-profiler==0.4.1
torchvision==0.14.1
tqdm==4.64.1
pytest
google-api-core
google-auth
google-auth-oauthlib
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
